Added EmailServer.py
The email server can currently send an email to our gmail from our gmail when called.  I have example code for a Python and Java VM thing that lets us call a Java method inside of Python.  The idea is that I take an input of the user's email address.  Using that we call the Java method to get the password and send the password to the user over SMTP in Python.  I also need to find a way to call this script from a webpage.